C5查询入库第二版

+++ yudao-module-steam/yudao-module-steam-biz/src/main/java/cn/iocoder/yudao/module/steam/dal/dataobject/othertemplate/OtherTemplateDO.java
@@ -54,7 +54,7 @@ public class OtherTemplateDO extends BaseDO {
     /**
      * 自动发货在售最低价
      */
-    private Double autoDeliverPrice;
+    private Float autoDeliverPrice;
     /**
      * 品质
      */
